gulp consumerPWA to generate consumer manifest and assets
